Move JSON decode utilities
